requirements.txt
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
asyncclick==8.1.3.4
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
python-dateutil==2.8.2
python-kasa==0.5.1
pyzmq==25.0.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
spotipy==2.22.1
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.15.0
